// frontend/src/app/models/band-links/band-links.model.ts
export interface BandLinks {
    social: {
      facebook?: string;
      instagram?: string;
      twitter?: string;
      youtube?: string;
      tiktok?: string;
    };
    streaming: {
      spotify?: string;
      appleMusic?: string;
      soundcloud?: string;
      bandcamp?: string;
      youtubeMusic?: string;
    };
  }
  